1.2
- Bug fixes
- Renamed Project
- Changed ReadMe

+++ division.cpp
@@ -9,7 +9,7 @@ std::string division()
     std::cin >> value1;
     std::cout << "Enter the second value for the division\n";
     std::cin >> value2;
-    float result{value1 / value2};
+    double result{value1 / value2};
     std::string result_str{std::to_string(result)};
     return "The result is " + result_str;
 }
